clarify the dusts-v2 proposal
based on what we learned so far, we decided to split the smoke tests and
routability tests into different sections. we also removed some of the jobs
that don't make sense in the routability test or the smoke tests. For example
the routability test doesn't make any assertion on the availibility of the ssh
proxy, therefore it was removed. Similarly, we don't need more than one Cell in
the smoke tests.

Signed-off-by: John Shahid <jvshahid@gmail.com>